Addressing code review comments #2 (2018-05-15).
